it actually works now. mangler bare tests og cleanup

+++ Main_remove_feriekasse.py
@@ -5,9 +5,12 @@ def resetFeriekasse():
     while (prompt != "y" and prompt != "n"):
         prompt = input("are you sure you want to reset the feriekasse? (y/n) ")
     if (prompt == "y"):
-        os.remove("Feriekasse.xlsx")
-        os.remove("latestMatchCovered.json")
-        os.remove("leaguesAndTeams.json")
+        if os.path.exists("Feriekasse.xlsx"):
+            os.remove("Feriekasse.xlsx")
+        if os.path.exists("./logs/leaguesAndTeams.json"):
+            os.remove("./logs/leaguesAndTeams.json")
+        if os.path.exists("./logs/latestMatchCovered.json"):
+            os.remove("./logs/latestMatchCovered.json") 
         print("... files removed")
     input("press enter to continue ...")
